+++ batcher/src/plugin_settings.py
@@ -183,7 +183,6 @@ def create_settings_for_convert():
   _set_sensitive_for_image_name_pattern_in_export_for_default_export_procedure(settings['main'])
   _set_file_extension_options_for_default_export_procedure(settings['main'])
 
-  settings['main/procedures'].connect_event('after-add-action', _on_after_add_align_procedure)
   settings['main/procedures'].connect_event('after-add-action', _on_after_add_export_procedure)
   settings['main/procedures'].connect_event('after-add-action', _on_after_add_scale_procedure)
 
@@ -332,7 +331,6 @@ def create_settings_for_export_images():
   _set_sensitive_for_image_name_pattern_in_export_for_default_export_procedure(settings['main'])
   _set_file_extension_options_for_default_export_procedure(settings['main'])
 
-  settings['main/procedures'].connect_event('after-add-action', _on_after_add_align_procedure)
   settings['main/procedures'].connect_event('after-add-action', _on_after_add_export_procedure)
   settings['main/procedures'].connect_event('after-add-action', _on_after_add_scale_procedure)
 
@@ -467,7 +465,6 @@ def create_settings_for_export_layers():
   _set_sensitive_for_image_name_pattern_in_export_for_default_export_procedure(settings['main'])
   _set_file_extension_options_for_default_export_procedure(settings['main'])
 
-  settings['main/procedures'].connect_event('after-add-action', _on_after_add_align_procedure)
   settings['main/procedures'].connect_event('after-add-action', _on_after_add_export_procedure)
   settings['main/procedures'].connect_event('after-add-action', _on_after_add_scale_procedure)
 
@@ -565,7 +562,6 @@ def create_settings_for_edit_layers():
       ]),
   ])
 
-  settings['main/procedures'].connect_event('after-add-action', _on_after_add_align_procedure)
   settings['main/procedures'].connect_event('after-add-action', _on_after_add_export_procedure)
   settings['main/procedures'].connect_event('after-add-action', _on_after_add_scale_procedure)
 
@@ -771,24 +767,6 @@ def _set_file_extension_options_for_default_export_procedure(main_settings):
     'after-reset',
     _set_file_format_export_options_from_extension,
     main_settings['file_extension'])
-
-
-def _on_after_add_align_procedure(_procedures, procedure, _orig_procedure_dict):
-  if procedure['orig_name'].value == 'align_and_offset_layers':
-    _set_sensitive_for_reference_layer_in_align(
-      procedure['arguments/reference_object'],
-      procedure['arguments/reference_layer'])
-
-    procedure['arguments/reference_object'].connect_event(
-      'value-changed',
-      _set_sensitive_for_reference_layer_in_align,
-      procedure['arguments/reference_layer'])
-
-
-def _set_sensitive_for_reference_layer_in_align(
-      reference_object_setting, reference_layer_setting):
-  reference_layer_setting.gui.set_sensitive(
-    reference_object_setting.value == builtin_procedures.AlignReferenceObjects.LAYER)
 
 
 def _on_after_add_export_procedure(_procedures, procedure, _orig_procedure_dict):